+++ Products/GenericSetup/tests/test_registry.py
@@ -1014,6 +1014,8 @@ class ProfileRegistryTests( BaseRegistryTests
                           , ConformsToIProfileRegistry
                           ):
 
+
+
     def _getTargetClass( self ):
 
         from Products.GenericSetup.registry import ProfileRegistry
@@ -1045,7 +1047,7 @@ class ProfileRegistryTests( BaseRegistryTests
                                 , PRODUCT
                                 , PROFILE_TYPE
                                 )
-
+        
         self.assertEqual( len( registry.listProfiles() ), 1 )
         self.assertEqual( len( registry.listProfileInfo() ), 1 )