+++ Products/GenericSetup/zcml.py
@@ -68,7 +68,6 @@ class IRegisterProfileDirective(Interface):
         required=False)
 
 
-_profile_regs = []
 def registerProfile(_context, name=u'default', title=None, description=None,
                     directory=None, provides=BASE, for_=None):
     """ Add a new profile to the registry.
@@ -82,8 +81,6 @@ def registerProfile(_context, name=u'default', title=None, description=None,
 
     if description is None:
         description = u''
-
-    _profile_regs.append('%s:%s' % (product, name))
 
     _context.action(
         discriminator=('registerProfile', product, name),
@@ -116,11 +113,8 @@ class IExportStepDirective(Interface):
         required=True)
 
 
-_export_step_regs = []
 
 def exportStep(context, name, handler, title=None, description=None):
-    global _export_step_regs
-    _export_step_regs.append(name)
 
     context.action(
         discriminator=('exportStep', name),
@@ -165,8 +159,6 @@ class IImportStepDependsDirective(Interface):
         required=True)
 
 
-_import_step_regs = []
-
 class importStep:
 
     def __init__(self, context, name, title, description, handler):
@@ -184,8 +176,6 @@ class importStep:
         self.dependencies += (name,)
 
     def __call__(self):
-        global _import_step_regs
-        _import_step_regs.append(self.name)
 
         self.context.action(
             discriminator=self.discriminator,
@@ -381,40 +371,14 @@ class upgradeSteps(object):
 #### cleanup
 
 def cleanUpProfiles():
-    global _profile_regs
-    for profile_id in _profile_regs:
-        try:
-            del _profile_registry._profile_info[profile_id]
-        except KeyError:
-            pass
-        try:
-            _profile_registry._profile_ids.remove(profile_id)
-        except ValueError:
-            pass
-
-    _profile_regs = []
     _upgrade_registry.clear()
 
 
 def cleanUpImportSteps():
-    global _import_step_regs
-    for name in _import_step_regs:
-        try:
-            _import_step_registry.unregisterStep(name)
-        except KeyError:
-            pass
-
-    _import_step_regs = []
+    pass
 
 def cleanUpExportSteps():
-    global _export_step_regs
-    for name in _export_step_regs:
-        try:
-            _export_step_registry.unregisterStep(name)
-        except KeyError:
-            pass
-
-    _export_step_regs = []
+    pass
 
 from zope.testing.cleanup import addCleanUp
 addCleanUp(cleanUpProfiles)